Add cmd line option that accepts /path/to/ini
Added a new command line option that accepts /path/to/ini/file as a
parameter. I also added a intial code for parsing the file to the
orchestrator. At the moment it just prints the file line by line.

I have also changed the get_opts() function a tiny bit so I could
check for parsing the command line args.