Move node.c from library into node dir and simplify setup
The node object will only ever be called from the node binary, so no
need to put it in the library.

We consolidated the option parsing into a main.c, and which just
creates an empty Node object, calls individual methods to set any
options that was passed and then calls start(), which does all
the actual work.

This also adds config file parsing, and support for a name option.

Signed-off-by: Alexander Larsson <alexl@redhat.com>