Redo directory and link setup
All the common code is now linked into the libhirte library which is
linked into all the tools instead of the shared sources being listed
multiple times and thus built multiple times.

All include paths are normalized to either:
  same binary => no include prefix
  common code => "libhirte/" include prefix
No relative include paths are used.

This also changes the library to be static library. It is large enough
to warrant sharing for code size reasons, and we definately are not
ready to support any shared library ABI.

Signed-off-by: Alexander Larsson <alexl@redhat.com>