orch: restructure orchestrator
This renames the Orchestrator object to Manager to match the new
proposed dbus api, and adds a ManagedNode object to use for tracked
nodes, which replaces the old peer-manager code.

So far, the managednode really doesn't do anything, that is to come.

Also, this disables the clang-analyzer-unix.Malloc lint check because
that doesn't seem to properly handle cleanup attributes.

Signed-off-by: Alexander Larsson <alexl@redhat.com>