Make nodes register their name with the manager and track lifetimes
When a node connects it must first register with its (unique) name,
we handle this and keep track of the name for later use.

We also track the lifetime of the node conncection, allowing a
later node to connect with the same name once the first disconnects.

I had to disable the cppcoreguidelines-interfaces-global-init lint, because
it seems incompatible with sd_bus vtable construction. Also I had to change
around the peer-bus support code a bit to make everythign work.

Signed-off-by: Alexander Larsson <alexl@redhat.com>