Add start of Node.ListUnits implementation
We accept the ListUnits call on the node object and forward it to
the agent on the node. That currently only replies with a
"not implemented" error, but that result is then returned
and then returned back to the original called.

The next step is to call the systemd method in the agent.

Signed-off-by: Alexander Larsson <alexl@redhat.com>