Add StartUnit operation
This calls StartUnit on the agent, and if that succeeds it signals
the availability of a new job object and returns its object path.

Eventually the agent will detect the matching JobRemoved signal from
systemd and will send a JobDone to the manager, which will then
emit a JobRemoved on the hirte job object.

Compared to the dbus API document some things are missing:
 * There is no queueing of conflicting operations
 * We don't support properties on the Job object.

Signed-off-by: Alexander Larsson <alexl@redhat.com>