Implement Subscribe/Unsubscribe on monitors
Each subscribe on a monitor is stored on that monitor, but also
forwarded to the manager which has a list of all subscriptions.

The manager then applies to each Node the subscriptions that apply to
it (i.e. all nodes for subscription with node "", or the named node
only for a named subscription).

Each Node keeps track of how many subscriptions there are for each
particular unit name, and whenever the count of a unit goes from 0 ->
1 or from 1 -> 0 we send a subscribe or unsubscribe to the agent (if
connected). In addition, when a new agent connects we send all outstanding
subscriptions to it.

The agent listens to all systemd unit property changes, and whenever
one is sent that matches a subscription we forward the change to the
manager.  The manager then finds all the monitors the change affects
and emits the UnitPropertiesChanged event on those monitors, ammending
the event data with the name of the node it happened on.

Signed-off-by: Alexander Larsson <alexl@redhat.com>