Monitor: Add virtual events for UnitNew and UnitRemoved
Regular systemd unit new/remove events are reported as before, but
with reason=`real`. Additionally, when a subscription is created
and we know that the unit is running already we report a unit-new
with type=`virual` (this can happen in either the agent, for the
first time we monitor the unit, or in the manager if we already
got a unit loaded reported from the agent.

We also report virtual unit-removed when an agent disconnects.

Signed-off-by: Alexander Larsson <alexl@redhat.com>