Switch argument order of GetUnitProperties
We now specify the unit first, which makes more sense. The order
goes in the order we address things, i.e. first the unit, then what
interface on the unit. This will make even more sense when
GetUnitProperty is added where the last argument is the actual property
on that interface.

Signed-off-by: Alexander Larsson <alexl@redhat.com>