Manager: Add virtual state changed to inactive on agent disconnect
If we previously reported a service as active (or activating, etc) and
then the agent disconnect, we now send a virtual a state change to
inactive before we send the virtual unit unload.

This makes it easier to keep track of the active state without having
to care about the loaded state.

Signed-off-by: Alexander Larsson <alexl@redhat.com>