Fix codespell check in CI
Currently codespell has been executed as a part of github CI, but it
didn't raise an error if some file had spelling errors (codespell just
fixed the spelling error). This issue is now fixed by adding
`check-codespell` target into Makefile, which raises an error, when a
spelling error is detected.

Signed-off-by: Martin Perina <mperina@redhat.com>