Add support for conf.d configuration files
This is a follow-up of #147 where it's suggested to use /etc/hirte/hirte.conf and /etc/hirte/agent.conf as configuration files, which could be modified by users to specify custom options.

But for automation it's not easy to parse current configuration file and save a new one with modified values, it would be much easier to use conf.d directory with custom user configuration files.

Here is the suggestion for the whole configuration files lineup for hirte manager:

    /usr/share/hirte/config/hirte-default.conf
        The base configuration file provided project defaults, which should not be modified by users
    /etc/hirte/hirte.conf
        The configurations file, where users can modified default values
    /etc/hirte/hirte.conf.d/
        Configuration directory where users might insert a configuration file (for example 99-custom-logging.conf)
        Only files with .conf suffix are used to load configuration changes
        Files in this directory are loaded in alphabetical order, so it's suggested to use number prefix to ease understanding of loading (for example 95-my-manager.conf and 99-custom-logging.conf)

Similar structure should be used also for hirte-agent:

    /usr/share/hirte-agent/config/hirte-default.conf
    /etc/hirte/agent.conf
    /etc/hirte/agent.conf.d/

Using above configuration files structure should also change the order of loading configurations files during hirte/hirte-agent startup:

    1)Load the default configuration values from /usr/share and store them to the configuration hash map
    2)Load user customizations from /etc/hirte/{hirte|agent}.conf and overwrite updated values in the configuration hash map
    3)Load user customizations files from /etc/hirte/{hirte|agent}.conf.d/ by alphabetical order and overwrite updated values in the configuration hash map
    4)Check if all required options have a valid value

Fixes: #148
Signed-off-by: Artiom Divak adivak@redhat.com
Signed-off-by: Artiom Divak <adivak@redhat.com>