Adjust the description of the project
Replace the use of the words "services orchestrator" by "services
controller". Orchestrator comes with a certain understanding and
expectations, especially around the management of containers, and we do
not want to set these expectations for hirte.
In addition, hirte isn't really an orchestrator by itself, it is just a
part of an architecture that would be responsible for managing services.

This commit thus adjusts the wording around the description of the
project.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>