Cleanup conf and man pages
Identify in the hire-agent man page where the /etc/hirte/agent.conf file
is supposed to be installed.

Also don't go over 120 characters in the conf files, to make them easier
to view in terminals as well as editors.

Add builtin default settings to conf files so users have better idea of
what is set when not commented out.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>