Better default dbus policy
The existing policy allowed anyone to send messages to hirte, which
isn't ideal. It also isn't a huge issue as the default policy
disallows sending any method calls even if you are allowed to send
messages in general.

However, it seems better to disallow messages in general, and just open
up a few specific methods. This change allows the basic introspection
but nothing more.

This is a bit more limited than what the systemd configuration uses, as that
allows anyone to use a lot of read-only APIs. However, readonly access to
hirte would be give read access to other machines too, so it makes sense to
be more limited.

Signed-off-by: Alexander Larsson <alexl@redhat.com>