Run lint and rpmbuild+e2e-tests in parallel
Originally we wanted lint to be executed before rpmbuild to catch simple
style error, but unfortunately markdown lint takes a lot of time and
also to peform C code linting we need to build the project due to
internal meson dependencies.
Current average times of job executions are:

 * lint:       2min 10sec
 * rpmbuild:   1min 13sec
 * e2e tests:  3min 20sec

So we can save some time in CI processing by running lint and rpmbuild
in parallel.

Signed-off-by: Martin Perina <mperina@redhat.com>