Merge e2e and rpmbuild into single workflow
In the previous patches lint jobs and unit test was moved from build
workflow. After this change build worfklow contains only rpm build and
execution of e2e workflow as a part of build workflow. This execution
was a bit confusing (especially in github UI when looking at workflow
execution results for a PR), so it would be better to join rpmbuild and
e2e tests into a single workflow.

Signed-off-by: Martin Perina <mperina@redhat.com>