agent: move the IPV4 and IPV6 validation
We should validate first if the data provided
is IPV4 or IPV6, if not, translate.

Additionally, when there is a host to be translated to IP we must assume ip_address can match IPV4 and IPV6 sizes. For such scenario let's make INET6_ADDRSTRLEN the default size as it can match both sizes.

from netinet/in.h:
INET6_ADDRSTRLEN = 46bytes
INET_ADDRSTRLEN = 16bytes

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>