Set Log Level API (#347)
* Set Log Level For The Manager

I added public API to the Manager.
Now a user  can use the with sdbus to set new log leve for the manger

Signed-off-by: Artiom Divak <adivak@redhat.com>

* Set Log Level For Node And Agent

The APIs of the node and the agent are connected.
The API of Agent is internal. So if the user wants
to change the log level for the agent he needs to
use the public API of the node

Signed-off-by: Artiom Divak <adivak@redhat.com>

* Set LogLevel from hirtectl

hirtectl will support set-loglevel command. This command is to set log
level on specific node or on the manager. For more info and usage
please go to
https://github.com/containers/hirte/issues/189#issuecomment-1560632309.

Signed-off-by: Artiom Divak <adivak@redhat.com>

---------

Signed-off-by: Artiom Divak <adivak@redhat.com>