fix segfault on disconnected anonymous node
When an anonymous node disconnects, it gets removed
from the manager. This already unsets the agent bus.
And after emitting state changing signals, the bus
and slots where unset again, causing a segfault. To
prevent this, call node_unset_agent_bus on disconnect
only for named nodes as this is already done when
removing the anonymous node from the manager.

Signed-off-by: Michael Engel <mengel@redhat.com>