provide error message when parsing config file fails
When parsing a config file fails, e.g. when a line
in the file is longer than the maximum allowed, inih
returns the line number as error code. Use this to
provide an error message to the user.
Also, set the option to abort parsing on the first
error occuring.

Signed-off-by: Michael Engel <mengel@redhat.com>