added python bindings to be installed from rpm for integration tests
Added the python3-bluechi package to be installed in
the local and snapshot container image and the required
dasbus dependency in the base image so that the bindings
can be used in the integration tests.
Since there is a breaking change in dasbas@v1.7 in the
naming of a data type, this import is wrapped with a
try so that both versions can be supported.

Signed-off-by: Michael Engel <mengel@redhat.com>