align return value for error on loading custom config file
For errors loading the default config file or any of the
confd files, either -EINVAL (result smaller than 0) or
the error code from the library is returned.
In the case of the custom config file this was not the case.
This commit aligns this and updates the affected unit test(s).

Signed-off-by: Michael Engel <mengel@redhat.com>