automate creating a GitHub release
The current release process of a new BlueChi version to GitHub
involves quite a few manual steps. Due to introducing submodules,
the default artifacts of a GitHub release does not contain the
respective source files - emphasizing the need for automation.
Based on the manual flow, its easy to automatically create a
release when a new tag is pushed.

Signed-off-by: Michael Engel <mengel@redhat.com>