adds a status and disconnected timestamp property to the agent interface
Fixes https://github.com/containers/bluechi/issues/543
In order to enable other applications on the managed node to react
to connection loss of the agent with the controller, add two new
properties: Status and DisconnectTimestamp
If the status will change, the agent will emit a property changed
signal on which other applications can listen. To provide information
for later when the disconnect happened, the timestamp property
can be queried.
This change also includes extending the D-Bus API description and
adding an integration test to it.

Signed-off-by: Michael Engel <mengel@redhat.com>