removed duplicate NodeConnectionStateChanged signal
Fixes https://github.com/containers/bluechi/issues/551
Since the property for the node status on the D-Bus API is marked
as emitting change (and signal is emitted), the signal of
NodeConnectionStateChanged signal is a duplicate and can be removed.
All usages of the NodeConnectionStateChanged signal in tests and
bluechictl are replaced by listening to the property change of
Status.

Signed-off-by: Michael Engel <mengel@redhat.com>