add tab feature to mkdocs
By using tabs, the code snippet for the api examples can
be grouped by use case and split by language.
Unfortunately, the pymdownx.tabbed seems to not work
properly with the readthedocs theme. Because of that the
theme was exchanged with material. The dependency for it
was added.
In order to sync tabs with the same title, e.g. multiple
tabs for python code, a dedicated js script was added.

Signed-off-by: Michael Engel <mengel@redhat.com>