tests: introduce running integration tests w/valgrind
In order to track memory leaks in various flows, adding an option to
run integration tests with valgrind. If environment variable
WITH_VALGRIND is set:
$ WITH_VALGRIND=1 tmt run -v
the bluechi-controller and bluechi-agent unit files will be configured
to run the binaries via valgrind, and then the process will collect
valgrind logs and will place them in the data directory of the test.
In addition, if any memory leaks would be detected - the test will
fail.

Signed-off-by: Mark Kemel <mkemel@redhat.com>