remove sudo from tests readme for base image creation
The tests README instructs the codeowners how to build base
images and push them to the registry. Specifically, the user
is instructed to run:

$ buildah login ...

and then:

$ sudo PUSH_MANIFEST=yes bash build-scripts...

This does not work, since buildah logs in for the current user, but
the base images are created by the build script for root user,
thus eventually the push action fails.

No need to run the build script as root, removing the `sudo`
statement

Signed-off-by: Mark Kemel <mkemel@redhat.com>