fix invalid read on disconnecting anonymous node
Fixes: https://github.com/containers/bluechi/issues/591

When an anonymous node has been disconnected in the controller
and the cleanup has been done, a log statement was made afterwards.
This statement differed based on the node name. Since the node
instance has been cleaned up (freed), but not set to NULL, valgrind
reported this as an invalid read. Moved the log statements
accordingly.

Signed-off-by: Michael Engel <mengel@redhat.com>