fix race condition for bluechi-proxy@.service
There is a potential race condition for ExecStart in
bluechi-proxy@.service when a stop signal is sent and
the ExecStart= command is still running. This leads
systemd to send a SIGTERM signal to the template service.
Therefore, SIGTERM needs to be an acceptable exit status,
otherwise the service transitions to a failed state.

Signed-off-by: Michael Engel <mengel@redhat.com>