log error when at least one list units request fails
Relates to:
https://github.com/eclipse-bluechi/bluechi/issues/567
https://github.com/eclipse-bluechi/bluechi/issues/568

When list units for all nodes is run and the requests are submitted,
it might happen that one of those is failing. Previously, this led
to a failure when reading the message for the node where the request
failed - which got propagated to the caller.
Now, an additional error log in BlueChi is made and the error code
is extracted from the sd_bus_message.
Also removes the top-most error print in the client to avoid printing
errors twice.

Signed-off-by: Michael Engel <mengel@redhat.com>