extended monitor API by adding and removing peers
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/187

Monitor signals are only sent to the creator of the monitor. The
new Peer API enables to add and remove additional listener on the
same monitor with its subscriptions.
A user with sufficient priviledges to make methods calls to BlueChi's
API can add and remove peers. The processes for the peers can be run
by a different user with no access to BlueChi's API, but will still
be able to receive events if added by a privileged process.

Signed-off-by: Michael Engel <mengel@redhat.com>