enable quicker tcp error notification
Fixes: https://github.com/eclipse-bluechi/bluechi/issues/652
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/648

When a connection between controller and agent is esablished and
is dropped later, e.g. removing the cable from the agent, the
disconnect is properly detected by the keepalive mechanism.
However, the keepalive takes a while to detect this (based on the
KEEPCNT of the system). If any command is issued during that time
frame, tcp will try to retransmit the data. Eventually, retransmitting
will be stopped and ICMP packets for host not reachable emitted.
The keepalive is not used then, which results in the connection
between agent and controller to be broken but not closed - a
disconnect is not detected.
By setting the IP_RECVERR socket option, errors such as the ICMP
host not reachable error will be delivered to the upper layer
(systemd event loop in our case) so that it can handle it.

Signed-off-by: Michael Engel <mengel@redhat.com>