added status property on manager interface
Fixes: https://github.com/eclipse-bluechi/bluechi/issues/603

Added a status property on the manager interface to express
the current status of the overall system - all nodes managed
by the controller. It also emits a signal if this state
changes, e.g. from online to degraded. Node (dis-)connects
that don't lead to a state change do not trigger this signal.

Signed-off-by: Michael Engel <mengel@redhat.com>