bluechictl: status command to show nodes list
In this commit we remove the `bluechictl monitor node-connection`
functionality and move it to `bluechictl status`, slightly expanding
it.

`bluechictl status` will print the list of nodes with their status
and when it was last seen.

`bluechictl status node` will print the same for for the node `node`

Also adding `-w/--watch` flag to `bluechictl status` command,
allowing to continuously monitor the node/nodes status (like the
original `bluechictl monitor node-connection`

Signed-off-by: Mark Kemel <mkemel@redhat.com>