enabled running integration tests on multipe hosts
Related to: https://github.com/eclipse-bluechi/bluechi/issues/662

In order to allow the integration tests being run on multiple hosts,
it is necessary to connect to them via SSH. This has been enabled by
moving the functions to create and get files as well as execute
commands to a client class which abstracts the communication to
either a container or a remote machine via SSH.
Since the host can either be a container or a remote machine, the
container class has been renamed to machine and implemented for
the controller and agent. The client used there can be set
dynamically.
Functions encapsulating systemctl or bluechictl commands were moved
to dedicated classes so that is clear which functionality is being
used.
The test class has also been changed so that either a container or
SSH setup is being used as this affects the setup as well as the
teardown. For example, a containerized setup can simply remove the
container while any added or changed files on a remote machine need
to be cleaned up properly.

Signed-off-by: Michael Engel <mengel@redhat.com>