skip creating coverage report when no info file is found
It can happen that provisioning, preparing or executing
the integration tests fails, e.g. failed to pull the
container image due to an infra outage. The finish step
of tmt, however, always gets executed. In order to avoid
failing the finish step, check for existing .info files
and skip if none are found.

Signed-off-by: Michael Engel <mengel@redhat.com>