Test agent connect via ControllerAddress
Added a test for agent to test if it connectes via ControllerAddress.
This test will run an agent and controller. The agent configuration will
be change to invalid ControllerAdress and then to valid one. The valid
configuration will get an IP the the controller is listenning to and the
invalid one will get an unlistened IP.

Related: https://github.com/eclipse-bluechi/bluechi/issues/668
Signed-off-by: Artiom Divak <adivak@redhat.com>