Add APIs to set controller host, port and address for Agent
This adds public APIs to set controller host, port and address to the
bluechi-agent.

Because it's possible to change the host, port and address used by agent
to connect controller on the fly using these APIs, the agent can
reconnect to controller of new node without restarting when the node
of working bluechi controller is changed.

If reconnect is true, try to reconnect to new bluechi-controller right
now.

Signed-off-by: Joonyoung Shim <joonyoung.shim@42dot.ai>