Fail the test if there is an error stopping bluechi services or gathering logs
Bluechi services are being stopped at the end of each test to be able to
collect logs or code coverage data. If bluechi service was stopped with
failure (for example due to core-dump), we need to fail the test.
Also if there is an error during logs gathering, we should fail the
test.