replaced pytest-timeout with custom timeout
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/772

When pytest-timeout triggers a timeout is triggered, the
test will be exited with pytest.fail and test artifacts
might not be collected - so debugging issues with timeouts
become hard to debug.
Therefore, replace pytest-timeout with a slim, custom
implementation that throws an error when timeout is reached.
In addition, lets distinguish between timeouts for setup,
test execution and artifact collection.

Signed-off-by: Michael Engel <mengel@redhat.com>