implemented job cancel
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/793

Implemented cancelling a job that has been created. Due to
possible race conditions, e.g. when the events of the job
being canceled are propagated and processed faster than
handling the job cancel callack, success is replied to the
caller as soon as the request to the agent has been submitted.
Errors only lead to logs in bluechi.

Signed-off-by: Michael Engel <mengel@redhat.com>