loosen selinux port policy
Fixes: https://github.com/eclipse-bluechi/bluechi/issues/836

The SELinux policy for BlueChi defined also the allowed port for
it during RPM install/uninstall/update. Due to the use of semanage
this required the policycoreutils-python-utils and pulled in python
as a dependency. To get rid of this dependency, we decided to
loosen the policy by removing the allowed port and hide the still
needed semanage for hirte migration behind a flag.
Users wanting to limit BlueChi to a certain port can still do that
by using semanage port on the kept bluechi_port_t type def.

Signed-off-by: Michael Engel <mengel@redhat.com>