Optimize LastSeenTimestamp property retrieval in controller
Previously, the last seen timestamp was recorded upon every reception
of a heartbeat signal.  This moves the logic to disconnected signal
handler so clock_gettime is called less often.  The property
getter (node_property_get_last_seen) is also modified to return the
current time if the node is online.

Signed-off-by: Daiki Ueno <dueno@redhat.com>