Switch packit integration tests to latest stable
Relates to: https://github.com/containers/podman/issues/22422

On fedora-rawhide the latest podman and composefs
version causes the `podman build` of the test container
to fail. Therefore, the tests are now using the
latest stable version of fedora as a workaround.
As soon as the related issue is resolved, a switch
back to rawhide can be done.

Signed-off-by: Michael Engel <mengel@redhat.com>