switch order of restoring created and changed files
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/879

When creating files, we first keep track of them in
the list for newly created ones. If a file is already
in that list and another file with the same path
(dir+name) is created we keep create a backup and
track it.
For clean-up we need to reverse that order - first
restore tracked backup files and then remove created
files. This ensures that there are no dangling files.

Signed-off-by: Michael Engel <mengel@redhat.com>