actively check if bluechi_machine_lib exists
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/879

In order to avoid failing mkdir calls on subsequent
tests for the bluechi_machine_lib, an active check is
being made. If it doesn't exist yet, we create the
directory and the library files. These files are not
tracked and thus not deleted after a test. This way,
we only create the lib once.

Signed-off-by: Michael Engel <mengel@redhat.com>