concatenate stderr and stdout in sshclient
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/879

The client for the podman container seems to
concatenate stderr and stdout in its output parameter.
In the SSH client stderr and stdout are different
return variables. In order to not break the `exec_run` API
of the Client class and still get a uniform behavior of
both clients, lets concat stderr and stdout in the SSH client.

Signed-off-by: Michael Engel <mengel@redhat.com>