Fix .conf suffix check for conf.d files
Fixes: https://github.com/eclipse-bluechi/bluechi/issues/892

The strstr command only checks if a substring exists within
another string. The found string does not need to be a
suffix.
Therefore, a dedicated ends_with function is implemented and
called for checking if the file name suffix is .conf. Unit
tests for this new function have been added as well.

Signed-off-by: Michael Engel <mengel@redhat.com>