Refactored ListUnitsRequest and ListUnitFilesRequest
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/889

The structs and functions used for ListUnits and ListUnitFiles are
identical for the most part. In order to reduce code duplication, the
implementation for both have been combined into a single struct with
unified functions. This should also allow for additional, similar features
in the future.

Signed-off-by: Michael Engel <mengel@redhat.com>