Move flush to bottom of printing unit
Fixes: https://github.com/eclipse-bluechi/bluechi/issues/946

Apparently, when directly printing after flushing stdout occasionally
results in the last print unit line for bluechictl status watch being
buffered. This leads to the bluechictl status watch integration test
being stuck and failing.
By switching the print and flush - so the flush is last - this seems
to be resolved.

Signed-off-by: Michael Engel <mengel@redhat.com>