Remove ANSI escape sequences from command outputs
Some commands, which are executed as a part of tests (for example
"systemctl is-enabled"), return output containing ANSI escape sequence.
Unfortunately tmt 1.37.0 started to check valid characters as a part of
junit report and those character caused failures. So this patch removes
ANSI escape sequences from logged data, so report in tmt 1.37.0 is
created successfully.

Signed-off-by: Martin Perina <mperina@redhat.com>