Emit ControllerAddress changed signal in SwitchController method
The changed signal for the agents ControllerAddress property
was emitted only in the agent_reconnect function when a
change in the address has been detected. This, however, led
to the agent first emitting a status disconnected signal
and then the address changed signal when the SwitchAddress
API has been called. The order of these signals should be
reversed.
The address changed signal is now also emitted inside the
SwitchController API function which changes the address.
In order to get the up-to-date value in the orch_addr field,
the new value is set here as well.
The ControllerAddress property changed signal will still be
emitted in the reconnect function to ensure any change in
the address due to name resolution is considered.

Signed-off-by: Michael Engel <mengel@redhat.com>