Added UDS support without sock activation
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/997

Due to the systemd socket unit for the bluechi-controller multiple
connection mechanisms such as TCP or UDS are supported. However, it
is not possible to start BlueChi directly (or the systemd service)
and use UDS since the FD for it gets passed from the socket unit.
In order to support this, another connection handler has been added
as well as the corresponding configuration option UseUDS, which is
set to false by default. A collision of the systemd socket unit and
the built-in UDS is being detected, logged and proceeded since the
desired event source is available. Additionally, a meson option has
been added to simplify using the same UDS path everywhere.

Signed-off-by: Michael Engel <mengel@redhat.com>