Disable bluechi-agent-user-bus tests
Relates to: https://github.com/eclipse-bluechi/bluechi/issues/900

The integration test for using the user bus with the bluechi-agent
fails on the GH CI as well now. Since passing integration tests is
required in order to merge, lets disable this test for now and
not block further development. Investigation needs to be done so
that we can fix this test and re-enable it.

Signed-off-by: Michael Engel <mengel@redhat.com>