controller: Allow configuring required selinux labels for node names
This extends how we configure allowed node names, and allows us to
configure per-node-name requirements.

A node name can now either be in the AllowedNodeNames list, or the
in a section named "[node FOO]", with a key Allowed=true. The later
is very useful from a .d dropin file.

If a named node is allowed, we look for the "RequiredSelinuxContext"
key in the above section, and if it is set, we require the connecting
agent to have the given selinux context type.

Note: We already disallow most processes (except bluechi_agent_t and
haproxy_t) to connect to the controller, but with this we can limit
what names they can claim. For example, we could enforce a special
name for an agent running in qm.