is-online system and agent shouldn't accept arguments
It was found that running
`bluechi-is-online agent --monitor 5000 --wait=10000`
Does not return error, although `--monitor` is a flag. The reason is
that `--monitor` is configured as a flag, and 5000 is treated as an
argument, which the method `agent` is configured to accept (min 0,
max 2), and the same goes for `system` method. Actually, these two
methods do not expect any arguments at all.

This change will limit the usage of `bluechi-is-online agent` and
`bluechi-is-online system` to not expect arguments, only the allowed
options, and `bluechi-is-online node` to accept only one argument

Signed-off-by: Mark Kemel <mkemel@redhat.com>