Change restart behavior in systemd units
The Restart=on-failure policy restarted the BlueChi components on any failure
5 times, thus triggering the max. restart limit (with a 100ms delay in between
each restart by default) which caused the service to fail in the end.
Since both, bluechi-agent and bluechi-controller, only exit with an error code
on an unrecoverable one, e.g. a configuration file has invalid syntax, these
restarts are unnecessary and can potentially cause issues. Therefore, the
Restart= option has been changed to on-abnormal so that BlueChi gets restarted
on unclean signals, timeouts, watchdog and OOM. See here for more details:
https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html#Restart=

Signed-off-by: Michael Engel <mengel@redhat.com>