Remove test data with too long log configs
The test cases for too long log configs (level, target and isquiet) causes
the bluechi-agent service to fail (as expected). However, this happens in
the test setup phase where we expect a running agent. Due to timing issues
this check might even pass, depending how fast the status is being checked.
This causes the test to be flacky.
Since such a long input for log configs is unrealistic and the proper level
for testing is on the unit level (test without the long data is already in
place), these cases have een removed.
In addition, the unused BluechiConfig variable for max line length has been
removed.

Signed-off-by: Michael Engel <mengel@redhat.com>